mnma-upload/requirements.txt
fastapi==0.111.0
uvicorn[standard]
PyYAML
boto3
pydantic
nest_asyncio
asyncio==3.4.3
fastapi-cognito
python-multipart
tenacity
psycopg2-binary
aiofiles>=23.0.0